Controller getSwitches() -> getSwitch(long dpid)
Necessary step for switch sync. We can't use
IFloodlightProvider.getSwitches() anymore. We now have a getSwitch()
method. There are also methods to get all dpids (getAllSwitchDpids())
and to get a map (getAllSwitchesMap()). These two return a snapshot.
TODO: need to check if they are used under the correct assumption in
existing code.

First stab at SwitchSyncRepresentation, which is the object I indent to
put in the SyncStore.