Add setup table default flows toggle
Setup table default flows is a part of switch handshake process. But
there are cases when we must not alter switch behaviour(set of installed
OF flows). To "cover" such scenarios new boolean option
"setupTablesDefaultFlows" was added for module
net.floodlightcontroller.core.internal.OFSwitchManager.

It it set to "YES" (default), FL will setup table's default rules on
transition to master state(same as current). If it set to "NO", FL will
not setup table's default flows.